Don't fail on file-level comments
The loader works great when there are comments within a selector, but
before this commit, if there were comments in a file that aren’t within
a selector, they were causing the loader to fail.

Such a comment doesn’t have any selectors, so this change simply checks
to see if `rule.selectors` is truth before continuing.